+++ UIKitSwaggerTests/Tests/ColorComponentTests.swift
@@ -381,4 +381,41 @@ class ColorComponentTests: XCTestCase {
         XCTAssertEqual(monochromeColor.white, CGFloat(0.88), "The white component of the monochrome color should be provided")
     }
 
+    //  MARK: Compoment conversion
+
+    func testRGBtoHSBConversion() {
+        rgbComponents = rgbColor.rgbComponents
+        hsbComponents = rgbComponents.asHSBComponents()
+        XCTAssertEqual(rgbColor.hsbComponents, hsbComponents, "The RGB components should be correctly converted into their HSB component equivalents")
+    }
+
+    func testRGBtoGrayscaleConversion() {
+        rgbComponents = rgbColor.rgbComponents
+        grayscaleComponents = rgbComponents.asGrayscaleComponents()
+        XCTAssertEqual(rgbColor.grayscaleComponents, grayscaleComponents, "The RGB components should be correctly converted into their grayscale component equivalents")
+    }
+
+    func testHSBtoRGBConversion() {
+        hsbComponents = hsbColor.hsbComponents
+        rgbComponents = hsbComponents.asRGBComponents()
+        XCTAssertEqual(hsbColor.rgbComponents, rgbComponents, "The HSB components should be correctly converted into their RGB component equivalents")
+    }
+
+    func testHSBtoGrayscaleConversion() {
+        hsbComponents = hsbColor.hsbComponents
+        grayscaleComponents = hsbComponents.asGrayscaleComponents()
+        XCTAssertEqual(hsbColor.grayscaleComponents, grayscaleComponents, "The HSB components should be correctly converted into their grayscale component equivalents")
+    }
+
+    func testGrayscaletoRGBConversion() {
+        grayscaleComponents = monochromeColor.grayscaleComponents
+        rgbComponents = grayscaleComponents.asRGBComponents()
+        XCTAssertEqual(monochromeColor.rgbComponents, rgbComponents, "The grayscale components should be correctly converted into their RGB component equivalents")
+    }
+
+    func testGrayscaletoHSBConversion() {
+        grayscaleComponents = monochromeColor.grayscaleComponents
+        hsbComponents = grayscaleComponents.asHSBComponents()
+        XCTAssertEqual(monochromeColor.hsbComponents, hsbComponents, "The grayscale components should be correctly converted into their HSB component equivalents")
+    }
 }